Update README.md
That is all happening in Spain so far, in terms of monetary fines (there has also been loads of reprimands). Because of the DPA's filling system, it is not possible to know the dates in which the fines are  published. I have been tracking them so I know the months in which they were imposed, which I have included in the list. Finally, LaLiga fine was infromally announced on June but was not published until July. Maybe worth changing it accordingly so the date criteria is the date of formal publication by the DPA. Best regards,